OCT-111: moved GetChannelLocalesQuery
